// src/shared/api/generated/models/aiOperationListItemDtoCategory.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type AiOperationListItemDtoCategory =
  (typeof AiOperationListItemDtoCategory)[keyof typeof AiOperationListItemDtoCategory];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AiOperationListItemDtoCategory = {
  CHAT: "CHAT",
  VOICE: "VOICE",
  SPEECH: "SPEECH",
  MEMORY: "MEMORY",
  AI_REVIEW: "AI_REVIEW",
  AI_ANALYSIS: "AI_ANALYSIS",
  CASE_INTELLIGENCE: "CASE_INTELLIGENCE",
  PROJECT_OVERHEAD: "PROJECT_OVERHEAD",
  CMS_AGENT: "CMS_AGENT",
} as const;