// src/shared/api/generated/models/aiOperationListItemDtoChargedAccount.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type AiOperationListItemDtoChargedAccount =
  (typeof AiOperationListItemDtoChargedAccount)[keyof typeof AiOperationListItemDtoChargedAccount];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AiOperationListItemDtoChargedAccount = {
  END_USER_ALLOWANCE: "END_USER_ALLOWANCE",
  PROJECT_BUDGET: "PROJECT_BUDGET",
  PROJECT_OVERHEAD: "PROJECT_OVERHEAD",
} as const;