// src/shared/api/generated/models/aiOperationListItemDtoStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type AiOperationListItemDtoStatus =
  (typeof AiOperationListItemDtoStatus)[keyof typeof AiOperationListItemDtoStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AiOperationListItemDtoStatus = {
  STARTED: "STARTED",
  RUNNING: "RUNNING",
  SUCCEEDED: "SUCCEEDED",
  FAILED: "FAILED",
  CANCELLED: "CANCELLED",
} as const;